moves sockets files to own directory, related #311
