+++ server/server.js
@@ -2,7 +2,7 @@ const hapi = require('hapi');
 const inert = require('inert');
 const fs = require('fs');
 const cookieAuth = require('hapi-auth-cookie');
-const createSocket = require('./sockets.js');
+const createSocket = require('./sockets/sockets.js');
 
 require('env2')('./config.env');
